Initialize AkkaCrawl Project
Create akkacrawl project as a java exercie to learn
how gradle works. AkkaCrawl leverage the java akka
library to work as actor mode. Currently, CrawlActor
DownloadActor, TerminatorActor and ExportActor are
added. Actor modle work with httpasyncclient library
to fetch, download, parse the html site. Exportor is
a module to generate the target files which make the
result easy to understand.